+++ tests/test_commands.py
@@ -1,5 +1,6 @@
 from io import StringIO
 
+import pytest
 from django.contrib.auth import get_user_model
 from django.contrib.auth.hashers import check_password
 from django.core.management import call_command
@@ -7,6 +8,8 @@ from django.core.management.base import CommandError
 from django.test import TestCase
 
 from oauth2_provider.models import get_application_model
+
+from . import presets
 
 
 Application = get_application_model()
@@ -111,6 +114,20 @@ class CreateApplicationTest(TestCase):
         app = Application.objects.get()
 
         self.assertEqual(app.user, user)
+
+    @pytest.mark.usefixtures("oauth2_settings")
+    @pytest.mark.oauth2_settings(presets.OIDC_SETTINGS_RW)
+    def test_application_created_with_algorithm(self):
+        call_command(
+            "createapplication",
+            "confidential",
+            "authorization-code",
+            "--redirect-uris=http://example.com http://example2.com",
+            "--algorithm=RS256",
+        )
+        app = Application.objects.get()
+
+        self.assertEqual(app.algorithm, "RS256")
 
     def test_validation_failed_message(self):
         output = StringIO()